# fastapi_backend/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "title": "北京健康评估系统API",
    "description": "北京健康评估系统的后端API接口",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/results/": {
      "get": {
        "tags": ["结果管理"],
        "summary": "获取结果列表",
        "operationId": "read_results",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "skip",
            "in": "query",
            "description": "跳过的记录数",
            "schema": {
              "type": "integer",
              "default": 0
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "返回的最大记录数",
            "schema": {
              "type": "integer",
              "default": 100
            }
          },
          {
            "name": "personnel_id",
            "in": "query",
            "description": "按人员ID筛选",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "personnel_name",
            "in": "query",
            "description": "按人员名称筛选",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取结果列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Result"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          }
        }
      }
    },
    "/api/results/{result_id}": {
      "get": {
        "tags": ["结果管理"],
        "summary": "获取特定结果",
        "operationId": "read_result",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取结果",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Result"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "结果不存在"
          }
        }
      },
      "delete": {
        "tags": ["结果管理"],
        "summary": "删除结果",
        "operationId": "delete_result",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功删除结果"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "结果不存在"
          }
        }
      }
    },
    "/api/results/{result_id}/report": {
      "get": {
        "tags": ["结果管理"],
        "summary": "获取结果报告",
        "operationId": "read_result_report",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "result_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取结果报告",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "report_content": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "结果或报告不存在"
          }
        }
      }
    },
    "/api/roles/": {
      "post": {
        "tags": ["角色管理"],
        "summary": "创建新角色",
        "operationId": "create_role",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "角色创建成功",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "角色名已存在"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      },
      "get": {
        "tags": ["角色管理"],
        "summary": "获取角色列表",
        "operationId": "read_roles",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取角色列表",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Role"
                  }
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          }
        }
      }
    },
    "/api/roles/{role_id}": {
      "get": {
        "tags": ["角色管理"],
        "summary": "获取特定角色",
        "operationId": "read_role",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "成功获取角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "角色不存在"
          }
        }
      },
      "put": {
        "tags": ["角色管理"],
        "summary": "更新角色",
        "operationId": "update_role",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RoleUpdate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功更新角色",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "角色名已存在"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "角色不存在"
          }
        }
      },
      "delete": {
        "tags": ["角色管理"],
        "summary": "删除角色",
        "operationId": "delete_role",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功删除角色"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "角色不存在"
          }
        }
      }
    },
    "/api/roles/{role_id}/permissions": {
      "post": {
        "tags": ["角色管理"],
        "summary": "为角色添加权限",
        "operationId": "add_permission_to_role",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RolePermissionCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "成功添加权限",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Role"
                }
              }
            }
          },
          "400": {
            "description": "权限已存在"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "角色或权限不存在"
          }
        }
      }
    },
    "/api/roles/{role_id}/permissions/{permission_id}": {
      "delete": {
        "tags": ["角色管理"],
        "summary": "从角色中移除权限",
        "operationId": "remove_permission_from_role",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "role_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          },
          {
            "name": "permission_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "成功移除权限"
          },
          "401": {
            "description": "未经授权"
          },
          "403": {
            "description": "权限不足"
          },
          "404": {
            "description": "角色、权限或角色权限关系不存在"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Role": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "permissions": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Permission"
            }
          }
        }
      },
      "RoleCreate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        },
        "required": ["name"]
      },
      "RoleUpdate": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "Permission": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          }
        }
      },
      "RolePermissionCreate": {
        "type": "object",
        "properties": {
          "permission_id": {
            "type": "integer"
          }
        },
        "required": ["permission_id"]
      },
...
      "Result": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer"
          },
          "data_id": {
            "type": "integer"
          },
          "personnel_id": {
            "type": "string"
          },
          "personnel_name": {
            "type": "string"
          },
          "evaluate_time": {
            "type": "string",
            "format": "date-time"
          },
          "evaluate_user": {
            "type": "string"
          },
          "stress_score": {
            "type": "number",
            "format": "float"
          },
          "depression_score": {
            "type": "number",
            "format": "float"
          },
          "anxiety_score": {
            "type": "number",
            "format": "float"
          },
          "social_score": {
            "type": "number",
            "format": "float"
          },
          "report_path": {
            "type": "string"
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
} 